+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 
 - New `influxdb` metrics target.
 - The `bloblang` input field `interval` now supports cron expressions.
-- New experimental `aws_kinesis` input.
+- New beta `aws_kinesis` and `aws_sqs` inputs.
 - The `bool` bloblang method now supports a wider range of string values.
 - New `reject` output type for conditionally rejecting messages.
 - All Redis components now support clustering and fail-over patterns.
@@ -19,6 +19,11 @@ All notable changes to this project will be documented in this file.
 - Fixed a panic on startup when using `if` statements within a `workflow` branch request or response map.
 - The `meta` bloblang function error messages now include the name of the required value.
 - Config unit tests now report processor errors when checks fail.
+
+### Changed
+
+- The experimental `aws_s3` input is now marked as beta.
+- The beta `kinesis_balanced` input is now deprecated.
 
 ## 3.35.0 - 2020-12-07